Finalised a first version of the gigsl-wrapper lib
All types that are supported by both glib and gsl have been introduced
to the library and are now available as blocks, vectors and matrices.
However, several TODOs are still open and curently just collected
in the README.md.
Next steps is testing the practicability of the lib in a live-example.
Than adding the basic infrastructure for the library, completing all the
TODOs and finally provide a release.